+++ www/cookie-handler.js
@@ -10,6 +10,7 @@ var cookieJar = new ToughCookie.CookieJar(store);
 
 module.exports = {
     setCookieFromString: setCookieFromString,
+    setCookie: setCookie,
     getCookieString: getCookieString,
     clearCookies: clearCookies,
     removeCookies: removeCookies
@@ -43,6 +44,12 @@ function setCookieFromString(url, cookieStr) {
     for (var i = 0; i < cookies.length; ++i) {
         cookieJar.setCookieSync(cookies[i], url, { ignoreError: true });
     }
+}
+
+function setCookie(url, cookie, options) {
+  options = options || {};
+  options.ignoreError = false;
+  cookieJar.setCookieSync(cookie, url, options);
 }
 
 function getCookieString(url) {